replace code in select.ts mouseDown and mouseUp with the startSelecting and finishSelecting utils
