+++ packages/ketcher-react/src/script/editor/tool/select.ts
@@ -16,28 +16,21 @@
 
 import {
   Action,
-  SGroup,
   fromAtomsAttrs,
   fromBondsAttrs,
-  fromItemsFuse,
-  fromMultipleMove,
   fromTextDeletion,
   fromTextUpdating,
-  getHoverToFuse,
-  getItemsToFuse,
-  FunctionalGroup,
-  fromSimpleObjectResizing,
-  fromArrowResizing,
-  ReStruct,
-  ReSGroup
+  FunctionalGroup
 } from 'ketcher-core'
 
 import LassoHelper from './helper/lasso'
-import { atomLongtapEvent } from './atom'
 import SGroupTool from './sgroup'
-import utils from '../shared/utils'
 import { xor } from 'lodash/fp'
 import { Editor } from '../Editor'
+import { startSelecting } from '../utils/selectItems/startSelecting'
+import { finishSelecting } from '../utils/selectItems/finishSelecting'
+import { continueSelecting } from '../utils/selectItems/continueSelecting'
+import { cancelSelecting } from '../utils/selectItems/cancelSelecting'
 
 class SelectTool {
   #mode: string
@@ -56,363 +49,15 @@ class SelectTool {
   }
 
   mousedown(event) {
-    const render = this.editor.render
-    const ctab = render.ctab
-    const molecule = ctab.molecule
-    const functionalGroups = molecule.functionalGroups
-    const selectedSgroups: any[] = []
-    const newSelected = { atoms: [] as any[], bonds: [] as any[] }
-    let actualSgroupId
-
-    this.editor.hover(null) // TODO review hovering for touch devicess
-
-    const selectFragment = this.#lassoHelper.fragment || event.ctrlKey
-    const ci = this.editor.findItem(
-      event,
-      selectFragment
-        ? [
-            'frags',
-            'sgroups',
-            'functionalGroups',
-            'sgroupData',
-            'rgroups',
-            'rxnArrows',
-            'rxnPluses',
-            'enhancedFlags',
-            'simpleObjects',
-            'texts'
-          ]
-        : [
-            'atoms',
-            'bonds',
-            'sgroups',
-            'functionalGroups',
-            'sgroupData',
-            'rgroups',
-            'rxnArrows',
-            'rxnPluses',
-            'enhancedFlags',
-            'simpleObjects',
-            'texts'
-          ],
-      null
-    )
-
-    if (ci && ci.map === 'atoms' && functionalGroups.size) {
-      const atomId = FunctionalGroup.atomsInFunctionalGroup(
-        functionalGroups,
-        ci.id
-      )
-      const atomFromStruct = atomId !== null && ctab.atoms.get(ci.id)?.a
-
-      if (atomFromStruct) {
-        for (const sgId of atomFromStruct.sgs.values()) {
-          actualSgroupId = sgId
-        }
-      }
-      if (
-        atomFromStruct &&
-        actualSgroupId !== undefined &&
-        !selectedSgroups.includes(actualSgroupId)
-      )
-        selectedSgroups.push(actualSgroupId)
-    }
-    if (ci && ci.map === 'bonds' && functionalGroups.size) {
-      const bondId = FunctionalGroup.bondsInFunctionalGroup(
-        molecule,
-        functionalGroups,
-        ci.id
-      )
-      const sGroupId = FunctionalGroup.findFunctionalGroupByBond(
-        molecule,
-        functionalGroups,
-        bondId
-      )
-      if (sGroupId !== null && !selectedSgroups.includes(sGroupId))
-        selectedSgroups.push(sGroupId)
-    }
-
-    if (selectedSgroups.length) {
-      for (const sgId of selectedSgroups) {
-        const sgroup = ctab.sgroups.get(sgId)
-        if (sgroup) {
-          const sgroupAtoms = SGroup.getAtoms(molecule, sgroup.item)
-          const sgroupBonds = SGroup.getBonds(molecule, sgroup.item)
-          newSelected.atoms.push(...sgroupAtoms) &&
-            newSelected.bonds.push(...sgroupBonds)
-        }
-      }
-      this.editor.selection(newSelected)
-    }
-
-    this.dragCtx = {
-      item: ci,
-      xy0: render.page2obj(event)
-    }
-
-    if (!ci || ci.map === 'atoms') {
-      atomLongtapEvent(this, render)
-    }
-
-    if (!ci) {
-      //  ci.type == 'Canvas'
-      if (!event.shiftKey) this.editor.selection(null)
-      delete this.dragCtx.item
-      if (!this.#lassoHelper.fragment) this.#lassoHelper.begin(event)
-      return true
-    }
-
-    let sel = closestToSel(ci)
-    const sgroups = ctab.sgroups.get(ci.id)
-    const selection = this.editor.selection()
-    if (ci.map === 'frags') {
-      const frag = ctab.frags.get(ci.id)
-      sel = {
-        atoms: frag.fragGetAtoms(ctab, ci.id),
-        bonds: frag.fragGetBonds(ctab, ci.id)
-      }
-    } else if (
-      (ci.map === 'sgroups' || ci.map === 'functionalGroups') &&
-      sgroups
-    ) {
-      const sgroup = sgroups.item
-      sel = {
-        atoms: SGroup.getAtoms(molecule, sgroup),
-        bonds: SGroup.getBonds(molecule, sgroup)
-      }
-    } else if (ci.map === 'rgroups') {
-      const rgroup = ctab.rgroups.get(ci.id)
-      sel = {
-        atoms: rgroup.getAtoms(render),
-        bonds: rgroup.getBonds(render)
-      }
-    } else if (ci.map === 'sgroupData') {
-      if (isSelected(selection, ci)) return true
-    }
-
-    if (event.shiftKey) {
-      this.editor.selection(selMerge(sel, selection, true))
-    } else {
-      this.editor.selection(null)
-      this.editor.selection(isSelected(selection, ci) ? selection : sel)
-    }
-    return true
+    startSelecting(event, this, this.#lassoHelper)
   }
 
   mousemove(event) {
-    const editor = this.editor
-    const render = editor.render
-    const restruct = editor.render.ctab
-    const dragCtx = this.dragCtx
-    if (dragCtx && dragCtx.stopTapping) dragCtx.stopTapping()
-    if (dragCtx && dragCtx.item) {
-      const atoms = restruct.molecule.atoms
-      const selection = editor.selection()
-      const shouldDisplayDegree =
-        dragCtx.item.map === 'atoms' &&
-        atoms?.get(dragCtx.item.id)?.neighbors.length === 1 &&
-        selection?.atoms?.length === 1 &&
-        !selection.bonds
-      if (shouldDisplayDegree) {
-        // moving selected objects
-        const pos = render.page2obj(event)
-        const angle = utils.calcAngle(dragCtx.xy0, pos)
-        const degrees = utils.degrees(angle)
-        this.editor.event.message.dispatch({ info: degrees + 'º' })
-      }
-      if (dragCtx.item.map === 'simpleObjects' && dragCtx.item.ref) {
-        if (dragCtx?.action) dragCtx.action.perform(render.ctab)
-        const current = render.page2obj(event)
-        const diff = current.sub(this.dragCtx.xy0)
-        dragCtx.action = fromSimpleObjectResizing(
-          render.ctab,
-          dragCtx.item.id,
-          diff,
-          current,
-          dragCtx.item.ref,
-          event.shiftKey
-        )
-        editor.update(dragCtx.action, true)
-        return true
-      }
-      if (dragCtx.item.map === 'rxnArrows' && dragCtx.item.ref) {
-        if (dragCtx?.action) dragCtx.action.perform(render.ctab)
-        const current = render.page2obj(event)
-        const diff = current.sub(dragCtx.xy0)
-        dragCtx.previous = current
-        dragCtx.action = fromArrowResizing(
-          render.ctab,
-          dragCtx.item.id,
-          diff,
-          current,
-          dragCtx.item.ref
-        )
-        editor.update(dragCtx.action, true)
-        return true
-      }
-      if (dragCtx.action) {
-        dragCtx.action.perform(restruct)
-        // redraw the elements in unshifted position, lest the have different offset
-        editor.update(dragCtx.action, true)
-      }
-
-      const expSel = editor.explicitSelected()
-      dragCtx.action = fromMultipleMove(
-        restruct,
-        expSel,
-        editor.render.page2obj(event).sub(dragCtx.xy0)
-      )
-
-      dragCtx.mergeItems = getItemsToFuse(editor, expSel)
-      editor.hover(getHoverToFuse(dragCtx.mergeItems))
-
-      editor.update(dragCtx.action, true)
-      return true
-    }
-
-    if (this.#lassoHelper.running()) {
-      const sel = this.#lassoHelper.addPoint(event)
-      editor.selection(
-        !event.shiftKey ? sel : selMerge(sel, editor.selection(), false)
-      )
-      return true
-    }
-
-    const maps =
-      this.#lassoHelper.fragment || event.ctrlKey
-        ? [
-            'frags',
-            'sgroups',
-            'functionalGroups',
-            'sgroupData',
-            'rgroups',
-            'rxnArrows',
-            'rxnPluses',
-            'enhancedFlags',
-            'simpleObjects',
-            'texts'
-          ]
-        : [
-            'atoms',
-            'bonds',
-            'sgroups',
-            'functionalGroups',
-            'sgroupData',
-            'rgroups',
-            'rxnArrows',
-            'rxnPluses',
-            'enhancedFlags',
-            'simpleObjects',
-            'texts'
-          ]
-
-    editor.hover(editor.findItem(event, maps, null), null, event)
-
-    return true
+    continueSelecting(event, this, this.#lassoHelper)
   }
 
   mouseup(event) {
-    const editor = this.editor
-    const selected = editor.selection()
-    const struct = editor.render.ctab
-    const molecule = struct.molecule
-    const functionalGroups = molecule.functionalGroups
-    const selectedSgroups: any[] = []
-    const newSelected = { atoms: [] as any[], bonds: [] as any[] }
-    let actualSgroupId
-
-    if (selected && functionalGroups.size && selected.atoms) {
-      for (const atom of selected.atoms) {
-        const atomId = FunctionalGroup.atomsInFunctionalGroup(
-          functionalGroups,
-          atom
-        )
-        const atomFromStruct = atomId !== null && struct.atoms.get(atomId)?.a
-
-        if (atomFromStruct) {
-          for (const sgId of atomFromStruct.sgs.values()) {
-            actualSgroupId = sgId
-          }
-        }
-        if (
-          atomFromStruct &&
-          actualSgroupId !== undefined &&
-          !selectedSgroups.includes(actualSgroupId)
-        )
-          selectedSgroups.push(actualSgroupId)
-      }
-    }
-
-    if (selected && functionalGroups.size && selected.bonds) {
-      for (const atom of selected.bonds) {
-        const bondId = FunctionalGroup.bondsInFunctionalGroup(
-          molecule,
-          functionalGroups,
-          atom
-        )
-        const sGroupId = FunctionalGroup.findFunctionalGroupByBond(
-          molecule,
-          functionalGroups,
-          bondId
-        )
-        if (sGroupId !== null && !selectedSgroups.includes(sGroupId))
-          selectedSgroups.push(sGroupId)
-      }
-    }
-
-    if (selectedSgroups.length) {
-      for (const sgId of selectedSgroups) {
-        const sgroup = struct.sgroups.get(sgId)
-        if (sgroup) {
-          const sgroupAtoms = SGroup.getAtoms(molecule, sgroup.item)
-          const sgroupBonds = SGroup.getBonds(molecule, sgroup.item)
-          newSelected.atoms.push(...sgroupAtoms) &&
-            newSelected.bonds.push(...sgroupBonds)
-        }
-      }
-    }
-
-    const dragCtx = this.dragCtx
-
-    if (dragCtx && dragCtx.stopTapping) dragCtx.stopTapping()
-
-    const possibleSaltOrSolvent = struct.sgroups.get(actualSgroupId)
-    const isDraggingSaltOrSolventOnStructure = SGroup.isSaltOrSolvent(
-      possibleSaltOrSolvent?.item.data.name
-    )
-    if (
-      (isDraggingSaltOrSolventOnStructure ||
-        this.isDraggingStructureOnSaltOrSolvent(dragCtx, struct.sgroups)) &&
-      dragCtx
-    ) {
-      preventSaltAndSolventsMerge(struct, dragCtx, editor)
-      delete this.dragCtx
-      if (this.#lassoHelper.running()) {
-        this.selectElementsOnCanvas(newSelected, editor, event)
-      }
-      return true
-    }
-
-    if (dragCtx && dragCtx.item) {
-      dragCtx.action = dragCtx.action
-        ? fromItemsFuse(struct, dragCtx.mergeItems).mergeWith(dragCtx.action)
-        : fromItemsFuse(struct, dragCtx.mergeItems)
-
-      editor.hover(null)
-      if (dragCtx.mergeItems) editor.selection(null)
-      if (dragCtx.action.operations.length !== 0) editor.update(dragCtx.action)
-
-      delete this.dragCtx
-    } else if (this.#lassoHelper.running()) {
-      // TODO it catches more events than needed, to be re-factored
-      this.selectElementsOnCanvas(newSelected, editor, event)
-    } else if (this.#lassoHelper.fragment) {
-      if (!event.shiftKey) editor.selection(null)
-    }
-    editor.event.message.dispatch({
-      info: false
-    })
-    return true
+    finishSelecting(event, this, this.#lassoHelper)
   }
 
   dblclick(event) {
@@ -421,7 +66,7 @@ class SelectTool {
     const { molecule, sgroups } = struct
     const functionalGroups = molecule.functionalGroups
     const render = editor.render
-    const ci = editor.findItem(
+    const closestItem = editor.findItem(
       event,
       ['atoms', 'bonds', 'sgroups', 'functionalGroups', 'sgroupData', 'texts'],
       null
@@ -430,10 +75,10 @@ class SelectTool {
     const atomResult: any[] = []
     const bondResult: any[] = []
     const result: any[] = []
-    if (ci && functionalGroups && ci.map === 'atoms') {
+    if (closestItem && functionalGroups && closestItem.map === 'atoms') {
       const atomId = FunctionalGroup.atomsInFunctionalGroup(
         functionalGroups,
-        ci.id
+        closestItem.id
       )
       const atomFromStruct = atomId !== null && struct.atoms.get(atomId)?.a
       if (
@@ -448,11 +93,11 @@ class SelectTool {
       )
         atomResult.push(atomId)
     }
-    if (ci && functionalGroups && ci.map === 'bonds') {
+    if (closestItem && functionalGroups && closestItem.map === 'bonds') {
       const bondId = FunctionalGroup.bondsInFunctionalGroup(
         molecule,
         functionalGroups,
-        ci.id
+        closestItem.id
       )
       const bondFromStruct = bondId !== null && struct.bonds.get(bondId)?.b
       if (
@@ -493,13 +138,13 @@ class SelectTool {
       this.editor.event.removeFG.dispatch({ fgIds: result })
       return
     }
-    if (!ci) return true
+    if (!closestItem) return true
 
     const selection = this.editor.selection()
 
-    if (ci.map === 'atoms') {
+    if (closestItem.map === 'atoms') {
       const action = new Action()
-      const atom = molecule.atoms.get(ci.id)
+      const atom = molecule.atoms.get(closestItem.id)
       const ra = editor.event.elementEdit.dispatch(atom)
       if (selection?.atoms) {
         const selectionAtoms = selection.atoms
@@ -514,8 +159,8 @@ class SelectTool {
           })
           .catch(() => null)
       }
-    } else if (ci.map === 'bonds') {
-      const bond = render.ctab.bonds.get(ci.id)?.b
+    } else if (closestItem.map === 'bonds') {
+      const bond = render.ctab.bonds.get(closestItem.id)?.b
       const rb = editor.event.bondEdit.dispatch(bond)
 
       if (selection?.bonds) {
@@ -531,15 +176,17 @@ class SelectTool {
           .catch(() => null) // w/o changes
       }
     } else if (
-      (ci.map === 'sgroups' &&
-        !FunctionalGroup.isFunctionalGroup(molecule.sgroups.get(ci.id))) ||
-      ci.map === 'sgroupData'
+      (closestItem.map === 'sgroups' &&
+        !FunctionalGroup.isFunctionalGroup(
+          molecule.sgroups.get(closestItem.id)
+        )) ||
+      closestItem.map === 'sgroupData'
     ) {
-      editor.selection(closestToSel(ci))
-      SGroupTool.sgroupDialog(editor, ci.id, null)
-    } else if (ci.map === 'texts') {
-      editor.selection(closestToSel(ci))
-      const text = molecule.texts.get(ci.id)
+      editor.selection(closestToSel(closestItem))
+      SGroupTool.sgroupDialog(editor, closestItem.id, null)
+    } else if (closestItem.map === 'texts') {
+      editor.selection(closestToSel(closestItem))
+      const text = molecule.texts.get(closestItem.id)
       const dialog = editor.event.elementEdit.dispatch({
         ...text,
         type: 'text'
@@ -548,9 +195,9 @@ class SelectTool {
       dialog
         .then(({ content }) => {
           if (!content) {
-            editor.update(fromTextDeletion(struct, ci.id))
+            editor.update(fromTextDeletion(struct, closestItem.id))
           } else if (content !== text?.content) {
-            editor.update(fromTextUpdating(struct, ci.id, content))
+            editor.update(fromTextUpdating(struct, closestItem.id, content))
           }
         })
         .catch(() => null)
@@ -559,57 +206,13 @@ class SelectTool {
   }
 
   mouseleave(_) {
-    if (this.dragCtx && this.dragCtx.stopTapping) this.dragCtx.stopTapping()
-
-    if (this.dragCtx && this.dragCtx.action) {
-      const action = this.dragCtx.action
-      this.editor.update(action)
-    }
-    if (this.#lassoHelper.running())
-      this.editor.selection(this.#lassoHelper.end())
-
-    delete this.dragCtx
-
-    this.editor.hover(null)
-  }
-
-  selectElementsOnCanvas(
-    elements: { atoms: number[]; bonds: number[] },
-    editor: Editor,
-    event
-  ) {
-    const sel =
-      elements.atoms.length > 0
-        ? selMerge(this.#lassoHelper.end(), elements, false)
-        : this.#lassoHelper.end()
-    editor.selection(
-      !event.shiftKey ? sel : selMerge(sel, editor.selection(), false)
-    )
-  }
-
-  isDraggingStructureOnSaltOrSolvent(dragCtx, sgroups: Map<number, ReSGroup>) {
-    let isDraggingOnSaltOrSolventAtom
-    let isDraggingOnSaltOrSolventBond
-    if (dragCtx?.mergeItems) {
-      const mergeAtoms = Array.from(dragCtx.mergeItems.atoms.values())
-      const mergeBonds = Array.from(dragCtx.mergeItems.bonds.values())
-      const sgroupsOnCanvas = Array.from(sgroups.values()).map(
-        ({ item }) => item
-      )
-      isDraggingOnSaltOrSolventAtom = mergeAtoms.some((atomId) =>
-        SGroup.isAtomInSaltOrSolvent(atomId as number, sgroupsOnCanvas)
-      )
-      isDraggingOnSaltOrSolventBond = mergeBonds.some((bondId) =>
-        SGroup.isBondInSaltOrSolvent(bondId as number, sgroupsOnCanvas)
-      )
-    }
-    return isDraggingOnSaltOrSolventAtom || isDraggingOnSaltOrSolventBond
+    cancelSelecting(this, this.#lassoHelper)
   }
 }
 
-function closestToSel(ci) {
+function closestToSel(closestItem) {
   const res = {}
-  res[ci.map] = [ci.id]
+  res[closestItem.map] = [closestItem.id]
   return res
 }
 
@@ -624,36 +227,12 @@ export function selMerge(selection, add, reversible: boolean) {
   return selection
 }
 
-function isSelected(selection, item) {
-  return (
-    selection && selection[item.map] && selection[item.map].includes(item.id)
-  )
-}
-
 function uniqArray(dest, add, reversible: boolean) {
   return add.reduce((_, item) => {
     if (reversible) dest = xor(dest, [item])
     else if (!dest.includes(item)) dest.push(item)
     return dest
   }, [])
-}
-
-function preventSaltAndSolventsMerge(
-  struct: ReStruct,
-  dragCtx,
-  editor: Editor
-) {
-  const action = dragCtx.action
-    ? fromItemsFuse(struct, null).mergeWith(dragCtx.action)
-    : fromItemsFuse(struct, null)
-  editor.hover(null)
-  if (dragCtx.mergeItems) {
-    editor.selection(null)
-  }
-  editor.update(action)
-  editor.event.message.dispatch({
-    info: false
-  })
 }
 
 export default SelectTool
